code clean up and added snapshot testing
