+++ .eslintrc
@@ -1,14 +1,6 @@
 {
-  "extends": [
-    "eslint:all",
-    "plugin:react/all",
-    "airbnb",
-    "plugin:jsx-a11y/recommended",
-    "prettier",
-    "prettier/react"
-  ],
+  "extends": ["react-app", "airbnb", "plugin:jsx-a11y/recommended", "prettier", "prettier/react"],
   "plugins": ["jsx-a11y", "prettier", "react"],
-
   "rules": {
     "semi": 0,
     "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
